[BACKPORT 2026.1][PLAT-21007][PLAT-21749][PLAT-21824][PLAT-21894] PA Collector in YBA HA without collected data sync + enable by default
Summary:
Support PA Collector in YBA HA environment.
Make sure only PA COllector on active YBA collects data.
Make sure PA COllector configuraiton in YBA properly updated after promotion.
Also, sync PA COllector config tables during YBA HA sync.

Currently, PA collector is disabled by default.
We're making it enabled by default.
Also, this makes sure we have perfAdvisor section in yba-ctl.yml after we upgrade from the older YBA-I release.

Original diffs: https://phorge.dev.yugabyte.com/D56037, https://phorge.dev.yugabyte.com/D53570, https://phorge.dev.yugabyte.com/D56318, https://phorge.dev.yugabyte.com/D56598

Test Plan:
Deployed 2 YBA instances with PA Collector enabled.
Created universe on one instance and enabled PA collection with advanced observability.
Set up HA between the instances. Performed promotion of standby instance.
Make sure PA collection is enabled on the new active instance, old PA data is missing, but new PA data is collected successfully.
Promote the ex-active instance back to active.
Make sure PA collection is enabled on the new active instance, PA data has gap in the middle and new PA data is collected successfully.

Installed YBA-I 2.29.0.0-b40.
Upgraded to the custom built YBA-I.
Made sure PA Collector is installed and properly configured.
Made sure yba-ctl.yml has perfAdvisor section with all the defaults.
Reconfigured with PA collector disabled. Made sure it worked and PA Collector is not running.
Reconfigured with PA Collector enabled. Made sure PA Collector is re-enabled and working as expected.

Reviewers: aravind.ramarathinam, sarthak.naithani, dshubin, #yba-api-review!, anijhawan, nbhatia, nsingh

Reviewed By: sarthak.naithani

Subscribers: nikhil, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D56217

+++ managed/src/main/java/com/yugabyte/yw/controllers/PACollectorController.java
@@ -112,6 +112,10 @@ public class PACollectorController extends AuthenticatedController {
     if (collector.getUuid() != null) {
       throw new PlatformServiceException(BAD_REQUEST, "Can't create collector with uuid set");
     }
+    if (collector.isEmbedded()) {
+      throw new PlatformServiceException(
+          BAD_REQUEST, "Can't create embedded collector via API - it is managed by YBA");
+    }
     collector.setCustomerUUID(customerUUID);
     collector = perfAdvisorService.save(collector, false);
     auditService()
@@ -153,6 +157,14 @@ public class PACollectorController extends AuthenticatedController {
     }
     collector.setCustomerUUID(customerUUID);
     PACollector currentPlatform = perfAdvisorService.getOrBadRequest(customerUUID, paUUID);
+    if (currentPlatform.isEmbedded()) {
+      throw new PlatformServiceException(
+          BAD_REQUEST, "Can't edit embedded collector via API - it is managed by YBA");
+    }
+    if (collector.isEmbedded()) {
+      throw new PlatformServiceException(
+          BAD_REQUEST, "Can't set embedded flag via API - it is managed by YBA");
+    }
     collector = CommonUtils.unmaskObject(currentPlatform, collector);
     collector = perfAdvisorService.save(collector, force);
     auditService()
@@ -175,6 +187,14 @@ public class PACollectorController extends AuthenticatedController {
   public Result deletePACollector(
       UUID customerUUID, UUID collectorUUID, boolean force, Http.Request request) {
     Customer.getOrBadRequest(customerUUID);
+    PACollector collector = perfAdvisorService.getOrBadRequest(customerUUID, collectorUUID);
+    if (collector.isEmbedded()) {
+      // The embedded collector is fully owned by EmbeddedCollectorInitializer - deleting it
+      // via the API would just leave it half-configured until the next initializer tick
+      // recreates it. Force users to disable it via runtime config (yb.pa.url) instead.
+      throw new PlatformServiceException(
+          BAD_REQUEST, "Can't delete embedded collector via API - it is managed by YBA");
+    }
 
     perfAdvisorService.delete(customerUUID, collectorUUID, force);
     auditService()
